+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Pending Release
 
 #### New Features
 - Added property `ApplicationCommand.mention`.
+- Added property `Message.mention` (see updated).
 - Added struct `Markdown`. This contains all methods related to markdowns. The following has been added:
   - `Markdown.blockQuote()`
   - `Markdown.bold()`
@@ -25,6 +26,7 @@ Pending Release
 - Fixed property `Emoji.url` returning a .png URL even when the emoji was animated (.gif).
 
 #### Updated
+- Property `GuildChannel.jumpUrl` and `Message.jumpUrl` now returns a suppressed link. With the Discord app update, messages that contain channel/message links are displayed as mentions. This changes it so that `.jumpUrl` and `.mention` have separate use cases.
 - With the addition of `Markdown`, the following has been changed:
   - **(Breaking Change)** Function `clean()` was renamed to `escape` and is now under `Markdown.escape()`.
   - **(Breaking Change)** Function `formatTimestamp()` was renamed to `timestamp` and is now under `Markdown.timestamp()`.